secure-login/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
sqlalchemy==2.0.36
argon2-cffi==23.1.0
itsdangerous==2.2.0
python-dotenv==1.0.1
email-validator==2.2.0
aiofiles==24.1.0